% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burbujas.R
\name{burbujas}
\alias{burbujas}
\title{Grafico de burbujas}
\usage{
burbujas(data, x, y, size, color, title, xlabel, ylabel)
}
\arguments{
\item{data}{Conjunto de datos con todas las variables que se ingresaran al grafico y su descripcion. Esta tabla debe ser en formato data table. No se ingresan las variables que no se quieren mostrar en ninguna opcion.}

\item{x}{Variable continua correspondiente al eje x del grafico de burbujas. Se ingresa indicando la tabla$variable, por ejemplo: data$variablex.}

\item{y}{Variable continua correspondiente al eje y del grafico de burbujas. Se ingresa indicando la tabla$variable, por ejemplo: data$variabley.}

\item{size}{Variable continua correspondiente al tamaño de cada burbuja. Se ingresa igualmente de manera tabla$variable_size.}

\item{color}{Variable categorica correspondiente al color de cada burbuja, la cual representa el grupo o categoria al que pertenece cada dato. Igualmente se ingresa de manera tabla$variable_color.}

\item{title}{Titulo del grafico. Ingresar como string.}
}
\value{
El grafico de burbujas.
}
\description{
Esta funcion presenta la informacion de multiples variables (continuas y categoricas) de manera grafica en el nivel mas desagregado, el cual es un grafico en 2 dimensiones con caracteristicas que permiten visualizar las distintas variables que se quieren mostrar.
}
\examples{
burbujas(data = iris, x = iris$Sepal.Length, y = iris$Petal.Length, size = iris$Sepal.Width,
         color = iris$Species, title = 'Iris', xlabel = 'Sepal Length', ylabel = 'Petal Length')
}
