% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tortas.R
\name{tortas}
\alias{tortas}
\title{Grafico de tortas}
\usage{
tortas(data)
}
\arguments{
\item{data}{Conjunto de datos con las variables categoricas a graficar. Esta tabla debe ser en formato data frame o data table. No ingresar variables que no sean correspondientes a categoricas del grafico (no ingresar ID ni continuas).}
}
\value{
Graficos de tortas de todas las variables categoricas ingresadas.
}
\description{
Resume la informacion de n variables (categoricas) en n graficos de torta, con el objetivo de ver las frecuencias de cada una de las categorias de cada variable.
}
\examples{
tortas(data.frame(HairEyeColor)[, c(1, 2, 3)])
}
